refactor : dto 패키지에 schedule하위 패키지 분리
